// src/types/business/BusinessAddressDetails.ts
export interface BusinessAddressDetails {
  id: number;
  userId: string;
  businessId: string;
  businessName: string;
  buildingName: string;
  floorNumber: string;
  street: string;
  city: string;
  country: string;
  county: string;
  postcode: string;
  latitude: number;
  longitude: number;
  createdAt: string;
  updatedAt: string;
}